fix 2d array passing

+++ multi_tile.cl
@@ -1,15 +1,15 @@
 __kernel void multi_tile(const int M, const int N, const int K,
-                      const __global float* A,
-                      const __global float* B,
-                      const __global float* Target,
-                      __global float* C) {
+                      const __global int* A,
+                      const __global int* B,
+                      const __global int* Target,
+                      __global int* C) {
     
     const int row = get_local_id(0); // Local row ID (max: arrA_GROUP_SIZE)
     const int col = get_local_id(1); // Local col ID (max: arrA_GROUP_SIZE)
     const int globalRow = arrA_GROUP_SIZE*get_group_id(0) + row; // Row ID of C (0..M)
     const int globalCol = arrA_GROUP_SIZE*get_group_id(1) + col; // Col ID of C (0..N)
 
-    float acc = 0.0f;
+    int acc = 0.0f;
     
     const int numTiles = K/arrA_GROUP_SIZE;
     for (int t=0; t<numTiles; t++) {
